refactored wizard controller features to new package
